fix adding repos
