+++ server/api/forges/[forge_id]/repos/add.post.ts
@@ -1,41 +1,4 @@
-import type { H3Event } from 'h3';
 import { repoSchema, userReposSchema } from '../../../../schemas';
-
-export async function getSessionHeader(event: H3Event) {
-  const config = useRuntimeConfig();
-
-  console.log(getHeaders(event));
-
-  const sessionName = config.auth.name || 'h3';
-
-  let sealedSession: string | undefined;
-
-  // Try header first
-  if (config.sessionHeader !== false) {
-    const headerName =
-      typeof config.sessionHeader === 'string'
-        ? config.sessionHeader.toLowerCase()
-        : `x-${sessionName.toLowerCase()}-session`;
-    const headerValue = event.node.req.headers[headerName];
-    if (typeof headerValue === 'string') {
-      sealedSession = headerValue;
-    }
-  }
-
-  // Fallback to cookies
-  if (!sealedSession) {
-    sealedSession = getCookie(event, sessionName);
-  }
-
-  if (!sealedSession) {
-    throw createError({
-      statusCode: 401,
-      statusMessage: 'Unauthorized',
-    });
-  }
-
-  return { [sessionName]: sealedSession };
-}
 
 export default defineEventHandler(async (event) => {
   const user = await requireUser(event);
@@ -51,9 +14,15 @@ export default defineEventHandler(async (event) => {
   const forgeId = parseInt(forgeIdFromParams, 10);
   const forge = await getUserForgeAPI(user, forgeId);
 
-  const { repoId } = (await readBody(event)) as { repoId: string };
+  const { remoteRepoId } = (await readBody(event)) as { remoteRepoId?: string };
+  if (!remoteRepoId) {
+    throw createError({
+      statusCode: 400,
+      statusMessage: 'remoteRepoId is required',
+    });
+  }
 
-  const forgeRepo = await forge.getRepo(repoId);
+  const forgeRepo = await forge.getRepo(remoteRepoId);
 
   const repo = await db
     .insert(repoSchema)
@@ -84,8 +53,7 @@ export default defineEventHandler(async (event) => {
     .run();
 
   const sessionHeader = await getSessionHeader(event);
-
-  await $fetch(`/api/repos/${repoId}/clone`, {
+  await $fetch(`/api/repos/${repo.id}/clone`, {
     method: 'POST',
     headers: {
       // forward session header
