+++ server/utils/auth.ts
@@ -95,3 +95,37 @@ export async function requireAccessToRepo(user: User, repoId: number) {
 
   return repo;
 }
+
+export async function getSessionHeader(event: H3Event) {
+  const config = useRuntimeConfig();
+
+  const sessionName = config.auth.name || 'h3';
+
+  let sealedSession: string | undefined;
+
+  // Try header first
+  if (config.sessionHeader !== false) {
+    const headerName =
+      typeof config.sessionHeader === 'string'
+        ? config.sessionHeader.toLowerCase()
+        : `x-${sessionName.toLowerCase()}-session`;
+    const headerValue = event.node.req.headers[headerName];
+    if (typeof headerValue === 'string') {
+      sealedSession = headerValue;
+    }
+  }
+
+  // Fallback to cookies
+  if (!sealedSession) {
+    sealedSession = getCookie(event, sessionName);
+  }
+
+  if (!sealedSession) {
+    throw createError({
+      statusCode: 401,
+      statusMessage: 'Unauthorized',
+    });
+  }
+
+  return { [`x-${sessionName.toLowerCase()}-session`]: sealedSession };
+}
